InitVar is now properly type-checked for dataclass
In Python dataclasses, InitVars are passed as arguments to __init__ and
__post_init__. Unfortunately, as of Python 3.7, InitVar does not pass
any type information regarding its contained type. For example, these is
not way to determine that InitVar[str] contains a str. InitVar is
currently defined as following in the Python standard library:

  class _InitVarMeta(type):
      def __getitem__(self, params):
          return self

  class InitVar(metaclass=_InitVarMeta):
      pass

As such, it is missing the __args__ attribute found on container classes
in the typing module. InitVars are passed as arguements to __init__ but
are not set there. Their type is InitVar, so we choose to simply ignore
their type (for now) and pass the value through __init__ if it is an
InitVar.

Because of this limitation, we rely on the __post_init__ method to
determine the type of an InitVar. We provide a check for __post_init__
method arguments. Any arguments passed to this method are assumed to be
InitVars defined for the data class. Their types can be annotated in
__post_init__, and the annotations are used to perform runtime type
checks on the InitVar.

Our treatment of InitVar could result in strange edge case bugs. This is
documented in the tests (see: test_known_bug_initvar_as_argument_passes)